PSY-422: split routes.go into per-domain files
routes.go was 1.1k lines containing every setupXxxRoutes function plus the
SetupRoutes entry point and rate-limit helpers. Split into ~30 per-domain
files in internal/api/routes/. Pure mechanical reorganization — no behavior
or registration-order changes.

routes.go is now 100 lines (entry point + ordered setup* calls).
shared.go holds RouteContext + rateLimitHandler + rateLimitUnlessAPIToken.
Inline auth registrations from the old SetupRoutes are factored into
setupProtectedAuthRoutes() in auth.go; SetupRoutes calls it in the same
position to preserve registration order.

Build: go build ./... clean.
Tests: go test -short ./... clean.
OpenAPI sanity check: /openapi.json output identical pre/post split
(diff'd via temporary dump test, removed before commit).

Closes PSY-422
